+++ Source/sspAddValue.h
@@ -31,7 +31,7 @@ public:
 	virtual ~SSpAddValue() {}
 
 	// Virtual methods
-	virtual float getValue() const override;
+	virtual double getValue() const override;
 	virtual bool verify(int& nErrors, int& nWarnings) const override;
 
 	// Accessors
