+++ Source/sspRandomValue.h
@@ -33,7 +33,7 @@ public:
 	virtual ~sspRandomValue() {}
 
 	// Virtual methods
-	virtual float getValue() const override;
+	virtual double getValue() const override;
 	virtual bool verify(int& nErrors, int& nWarnings) const override;
 
 	// Accessors
