+++ Source/sspTimeMap.cpp
@@ -23,8 +23,8 @@ void sspTimeMap::computeLinearFactors()
 {
 	auto inp_min = clock_min_.total_seconds();
 	auto inp_max = clock_max_.total_seconds();
-	lin_a_ = (inp_min < inp_max) ? (outp_max_ - outp_min_) / static_cast<float>(inp_max - inp_min) : 1.0f;
-	lin_b_ = outp_min_ - lin_a_ * static_cast<float>(clock_min_.total_seconds());
+	lin_a_ = (inp_min < inp_max) ? (outp_max_ - outp_min_) / static_cast<double>(inp_max - inp_min) : 1.0f;
+	lin_b_ = outp_min_ - lin_a_ * static_cast<double>(clock_min_.total_seconds());
 }
 
 
@@ -41,7 +41,7 @@ void sspTimeMap::setInputRange(const time_duration & min, const time_duration & 
 	computeLinearFactors();
 }
 
-void sspTimeMap::setOutputRange(float fMin, float fMax)
+void sspTimeMap::setOutputRange(double fMin, double fMax)
 {
 	// May be inverted
 	outp_min_ = fMin;
@@ -49,10 +49,10 @@ void sspTimeMap::setOutputRange(float fMin, float fMax)
 	computeLinearFactors();
 }
 
-float sspTimeMap::getValue() const
+double sspTimeMap::getValue() const
 {
 	auto now = second_clock::local_time().time_of_day();
-	return lin_a_ * static_cast<float>(now.total_seconds()) + lin_b_;
+	return lin_a_ * static_cast<double>(now.total_seconds()) + lin_b_;
 }
 
 bool sspTimeMap::verify(int & nErrors, int & nWarnings) const
@@ -65,7 +65,7 @@ bool sspTimeMap::verify(int & nErrors, int & nWarnings) const
 	if (clock_min_ == clock_max_) {
 		SSP_LOG_WRAPPER_ERROR(nErrors, bReturn) << getName() << ": time range has zero seconds";
 	}
-	if (std::abs(outp_max_ - outp_min_) < std::numeric_limits<float>::epsilon()) {
+	if (std::abs(outp_max_ - outp_min_) < std::numeric_limits<double>::epsilon()) {
 		SSP_LOG_WRAPPER_WARNING(nWarnings, bReturn) << getName() << ": output has zero range";
 	}
